fix(karma): don't add debug log-level on travis, let module decide
The debug log-level produces some very wordy output, which is usually not very helpful.

When not intending to debug the test run, this isn't very useful, so this should be
enabled at the discretion of the package being tested. It makes the watchtower.js test
runs look much scarier than is really necessary.